test: 흑돌 백돌 금수 테스트 추가
